переехал на jsx и scss, вернул изначальный вид сайту
