+++ src/App.jsx
@@ -1,9 +1,5 @@
 import React from 'react';
-import './App.css';
 import TodoList from './ToDoList/ToDoList';
-
-
-
 
 class App extends React.Component{
 	render(){
