+++ src/App.module.scss
@@ -36,3 +36,4 @@
     transform: rotate(360deg);
   }
 }
+
