--- src/ToDoList/ToDoList.css
@@ -1,11 +1,0 @@
-html{
-    background-color: #121212;
-    
-}
-  
-.Heading{
-    text-align: center;
-    margin-top: 10%;
-    color: white;
-    margin-bottom: 10%;
-}
